[virt] kvm: arm64: gicv4.1: Plumb SGI implementation selection in the distributor
Message-id: <20200514034601.64374-32-gshan@redhat.com>
Patchwork-id: 307479
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH v3 31/52] KVM: arm64: GICv4.1: Plumb SGI implementation selection in the distributor
Bugzilla: 1833572
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Marc Zyngier <maz@kernel.org>

The GICv4.1 architecture gives the hypervisor the option to let
the guest choose whether it wants the good old SGIs with an
active state, or the new, HW-based ones that do not have one.

For this, plumb the configuration of SGIs into the GICv3 MMIO
handling, present the GICD_TYPER2.nASSGIcap to the guest,
and handle the GICD_CTLR.nASSGIreq setting.

In order to be able to deal with the restore of a guest, also
apply the GICD_CTLR.nASSGIreq setting at first run so that we
can move the restored SGIs to the HW if that's what the guest
had selected in a previous life.

Signed-off-by: Marc Zyngier <maz@kernel.org>
Reviewed-by: Zenghui Yu <yuzenghui@huawei.com>
Link: https://lore.kernel.org/r/20200304203330.4967-21-maz@kernel.org
(cherry picked from commit 2291ff2f2a569b7b7a64aff3d05e1d51fe0fd66a)
Signed-off-by: Gavin Shan <gshan@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>