[virt] kvm: arm: vgic: Synchronize the whole guest on GIC{D, R}_I{S, C}ACTIVER read
Message-id: <20200514034601.64374-44-gshan@redhat.com>
Patchwork-id: 307490
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH v3 43/52] KVM: arm: vgic: Synchronize the whole guest on GIC{D, R}_I{S, C}ACTIVER read
Bugzilla: 1833572
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Marc Zyngier <maz@kernel.org>

When a guest tries to read the active state of its interrupts,
we currently just return whatever state we have in memory. This
means that if such an interrupt lives in a List Register on another
CPU, we fail to obsertve the latest active state for this interrupt.

In order to remedy this, stop all the other vcpus so that they exit
and we can observe the most recent value for the state. This is
similar to what we are doing for the write side of the same
registers, and results in new MMIO handlers for userspace (which
do not need to stop the guest, as it is supposed to be stopped
already).

Reported-by: Julien Grall <julien@xen.org>
Reviewed-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
(cherry picked from commit 9a50ebbffa9862db7604345f5fd763122b0f6fed)
Signed-off-by: Gavin Shan <gshan@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>