[kernel] padata: Remove broken queue flushing
Message-id: <E1jZ8hp-0001Zk-UK@fornost.hmeau.com>
Patchwork-id: 307538
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 14/23] padata: Remove broken queue flushing
Bugzilla: 1671674
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>

RHEL8.3 Bugzilla #1671674
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28537447

commit 07928d9bfc81640bab36f5190e8725894d93b659

The function padata_flush_queues is fundamentally broken because
it cannot force padata users to complete the request that is
underway.  IOW padata has to passively wait for the completion
of any outstanding work.

As it stands flushing is used in two places.  Its use in padata_stop
is simply unnecessary because nothing depends on the queues to
be flushed afterwards.

The other use in padata_replace is more substantial as we depend
on it to free the old pd structure.  This patch instead uses the
pd->refcnt to dynamically free the pd structure once all requests
are complete.

Fixes: 2b73b07ab8a4 ("padata: Flush the padata queues actively")
Cc: <stable@vger.kernel.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Reviewed-by: Daniel Jordan <daniel.m.jordan@oracle.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>