[kernel] perf/core: Optimize perf_install_in_event()
Message-id: <20200514101858.720-8-mpetlan@redhat.com>
Patchwork-id: 307569
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 007/300] perf/core: Optimize perf_install_in_event()
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit db0503e4f6751f2c719d002ba1becd1811633e6e
Author: Peter Zijlstra <peterz@infradead.org>
Date: Mon Oct 21 16:02:39 2019 +0200

description
===========
Andi reported that when creating a lot of events, a lot of time is
spent in IPIs and asked if it would be possible to elide some of that.

Now when, as for example the perf-tool always does, events are created
disabled, then these events will not need to be scheduled when added
to the context (they're still disable) and therefore the IPI is not
required -- except for the very first event, that will need to set
ctx->is_active.

( It might be possible to set ctx->is_active remotely for cpu_ctx, but
  we really need the IPI for task_ctx, so lets not make that
  distinction. )

Also use __perf_effective_state() since group events depend on the
state of the leader, if the leader is OFF, the whole group is OFF.

So when sibling events are created enabled (XXX check tool) then we
only need a single IPI to create and enable the whole group (+ that
initial IPI to initialize the context).

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>