[kernel] perf/core: Optimize perf_init_event()
Message-id: <20200514101858.720-9-mpetlan@redhat.com>
Patchwork-id: 307570
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 008/300] perf/core: Optimize perf_init_event()
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 66d258c5b048840991de49697264af75f5b09def
Author: Peter Zijlstra <peterz@infradead.org>
Date: Thu Oct 17 20:31:03 2019 +0200

description
===========
Andi reported that he was hitting the linear search in
perf_init_event() a lot. Make more agressive use of the IDR lookup to
avoid hitting the linear search.

With exception of PERF_TYPE_SOFTWARE (which relies on a hideous hack),
we can put everything in the IDR. On top of that, we can alias
TYPE_HARDWARE and TYPE_HW_CACHE to TYPE_RAW on the lookup side.

This greatly reduces the chances of hitting the linear search.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>