[tools] perf trace: Separate 'struct syscall_fmt' definition from syscall_fmts variable
Message-id: <20200514101858.720-34-mpetlan@redhat.com>
Patchwork-id: 307598
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 033/300] perf trace: Separate 'struct syscall_fmt' definition from syscall_fmts variable
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 9b2036cd329924082acfa5dec58deec12fa1f5e8
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Oct 1 15:16:33 2019 -0300

description
===========
As this has all the things needed to format tracepoints events, not just
syscalls, that, after all, are just tracepoints with a set in stone ABI,
i.e. order and number of parameters.

For tracepoints we'll create a

  static struct syscall_fmt tracepoint_fmts[]

array and will fill the ->arg[] entries with the beautifier for each
positional argument and record the name, then, when we need it, we'll
just check that the position has the same name, maybe even type, so that
we can do some check that the tracepoint hasn't changed, if it has, we
can even reorder things.

Keep calling it syscall_fmt but use it as well for tracepoints, do it
this way to minimize changes and reuse what is in place for syscalls,
we'll see.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>