[tools] perf trace: Add a strtoul() method to 'struct syscall_arg_fmt'
Message-id: <20200514101858.720-62-mpetlan@redhat.com>
Patchwork-id: 307623
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 061/300] perf trace: Add a strtoul() method to 'struct syscall_arg_fmt'
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 3f41b77843b338e836f52cc2d486be689d6cb9c1
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Oct 9 16:06:43 2019 -0300

description
===========
This will go from a string to a number, so that filter expressions can
be constructed with strings and then, before applying the tracepoint
filters (or eBPF, in the future) we can map those strings to numbers.

The first one will be for 'msr' tracepoint arguments, but real quickly
we will be able to reuse all strarrays for that.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>