[tools] perf tools: Use perf_mmap way to detect aux mmap
Message-id: <20200514101858.720-74-mpetlan@redhat.com>
Patchwork-id: 307635
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 073/300] perf tools: Use perf_mmap way to detect aux mmap
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 1d40ae4e1784bfa1646fd153ca022db21511284f
Author: Jiri Olsa <jolsa@kernel.org>
Date: Mon Oct 7 14:53:16 2019 +0200

description
===========
We will move this code to libperf shortly, so we need to free it of
'struct auxtrace_mmap' usage, because it won't be available in libperf
(for now).

The perf_event_mmap_page::aux_size is set when the aux mmap is mapped,
so the check is equivalent.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>