[tools] perf report: Add warning when libunwind not compiled in
Message-id: <20200514101858.720-99-mpetlan@redhat.com>
Patchwork-id: 307656
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 098/300] perf report: Add warning when libunwind not compiled in
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 800d3f561659b5436f8c57e7c26dd1f6928b5615
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Oct 11 10:21:22 2019 +0800

description
===========
We received a user report that call-graph DWARF mode was enabled in
'perf record' but 'perf report' didn't unwind the callstack correctly.
The reason was, libunwind was not compiled in.

We can use 'perf -vv' to check the compiled libraries but it would be
valuable to report a warning to user directly (especially valuable for
a perf newbie).

The warning is:

Warning:
Please install libunwind development packages during the perf build.

Both TUI and stdio are supported.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>