[tools] perf trace: Introduce accessors to trace specific evsel->priv
Message-id: <20200514101858.720-122-mpetlan@redhat.com>
Patchwork-id: 307679
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 121/300] perf trace: Introduce accessors to trace specific evsel->priv
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit fecd990720306f93151747771f16bca71bb29c33
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Oct 17 08:15:11 2019 -0300

description
===========
We're using evsel->priv in syscalls:sys_{enter,exit}_SYSCALL and in
raw_syscalls:sys_{enter,exit} to cache the offset of the common fields,
the multiplexor id/syscall_id in the sys_enter case and syscall_id + ret
for sys_exit.

And for the rest of the tracepoints we use it to have a syscall_arg_fmt
array to have scnprintf/strtoul for tracepoint args.

So we better clearly mark them with accessors so that we can move to
having a 'struct evsel_trace' struct for all 'perf trace' specific
evsel->priv usage.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>