[tools] perf parse: Ensure config and str in terms are unique
Message-id: <20200514101858.720-178-mpetlan@redhat.com>
Patchwork-id: 307736
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 177/300] perf parse: Ensure config and str in terms are unique
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit b6645a72359590ad7f57050d192cd5d8885320df
Author: Ian Rogers <irogers@google.com>
Date: Wed Oct 30 15:34:43 2019 -0700

description
===========
Make it easier to release memory associated with parse event terms by
duplicating the string for the config name and ensuring the val string
is a duplicate.

Currently the parser may memory leak terms and this is addressed in a
later patch.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>