[tools] perf annotate: Pass a 'map_symbol' in places receiving a pair of 'map' and 'symbol' pointers
Message-id: <20200514101858.720-199-mpetlan@redhat.com>
Patchwork-id: 307763
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 198/300] perf annotate: Pass a 'map_symbol' in places receiving a pair of 'map' and 'symbol' pointers
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 2975489458c59ce2e348b1b3aef5d8d2acb5cc8d
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Nov 4 11:10:00 2019 -0300

description
===========
We are already passing things like:

  symbol__annotate(ms->sym, ms->map, ...)

So shorten the signature of such functions to receive the 'map_symbol'
pointer.

This also paves the way to having the 'struct map_groups' pointer in the
'struct map_symbol' so that we can get rid of 'struct map'->groups.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>