[tools] pref tools: Make 'struct addr_map_symbol' contain 'struct map_symbol'
Message-id: <20200514101858.720-202-mpetlan@redhat.com>
Patchwork-id: 307766
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 201/300] pref tools: Make 'struct addr_map_symbol' contain 'struct map_symbol'
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit d46a4cdf49937b0b3abeb2cd7fa5dc65795e7ea7
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Nov 4 15:57:38 2019 -0300

description
===========
So that we pass that substructure around and with it consolidate lots of
functions that receive a (map, symbol) pair and now can receive just a
'struct map_symbol' pointer.

This further paves the way to add 'struct map_groups' to 'struct
map_symbol' so that we can have all we need for annotation so that we
can ditch 'struct map'->groups, i.e. have the map_groups pointer in a
more central place, avoiding the pointer in the 'struct map' that have
tons of instances.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>