[tools] perf record: No need to process the synthesized MMAP events twice
Message-id: <20200514101858.720-217-mpetlan@redhat.com>
Patchwork-id: 307780
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 216/300] perf record: No need to process the synthesized MMAP events twice
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 6e0a9b3dfaaf93476b34825e53c4ec065267081e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Nov 14 12:15:34 2019 -0300

description
===========
At the end of a 'perf record' session, by default, we'll process all
samples and populate the threads, maps, etc so as to find out which of
the DSOs got samples, to reduce the size of the build-id table we'll
add to the perf.data headers.

But we don't need to process the PERF_RECORD_MMAP events synthesized
for the kernel modules, as we have those already via
perf_session__create_kernel_maps(), so add mmap/mmap2 handlers that
first look at event->header.misc to see if the event is for a user map,
bailing out if not.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>