[tools] perf record: Add a function to test for kernel support for AUX area sampling
Message-id: <20200514101858.720-234-mpetlan@redhat.com>
Patchwork-id: 307797
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 233/300] perf record: Add a function to test for kernel support for AUX area sampling
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 9bca1a4ef5034f0a82861ac0375eb0272c5ce04e
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Nov 15 14:42:13 2019 +0200

description
===========
Architectures are expected to know if AUX area sampling is supported by
the hardware. Add a function perf_can_aux_sample() which will determine
whether the kernel supports it.

Committer notes:

I reported that this message was taking place on a kernel without the
required bits:

  # perf record --aux-sample -e '{intel_pt//u,branch-misses:u}'
  Error:
  The sys_perf_event_open() syscall returned with 7 (Argument list too long) for event (branch-misses:u).
  /bin/dmesg | grep -i perf may provide additional information.

Adrian sent a patch addressing it, with this explanation:

 ----
  perf_can_aux_sample_size() always returned true because it did not pass
  the attribute size to sys_perf_event_open, nor correctly check the
  return value and errno.
 ----

After applying it I get, later in the series, when --aux-sample is
added:

  # perf record --aux-sample -e '{intel_pt//u,branch-misses:u}'
  AUX area sampling is not supported by kernel

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>