[tools] perf pmu: When using default config, record which bits of config were changed by the user
Message-id: <20200514101858.720-243-mpetlan@redhat.com>
Patchwork-id: 307808
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 242/300] perf pmu: When using default config, record which bits of config were changed by the user
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit a1ac7de6902c1ea6def7a743f1d2e6ba429684b3
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Nov 15 14:42:22 2019 +0200

description
===========
Default config for a PMU is defined before selected events are parsed.
That allows the user-entered config to override the default config.

However that does not allow for changing the default config based on
other options.

For example, if the user chooses AUX area sampling mode, in the case of
Intel PT, the psb_period needs to be small for sampling, so there is a
need to set the default psb_period to 0 (2 KiB) in that case. However
that should not override a value set by the user. To allow for that,
when using default config, record which bits of config were changed by
the user.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>