[tools] perf machine: Fill map_symbol->maps in append_inlines() to fix segfault
Message-id: <20200514101858.720-280-mpetlan@redhat.com>
Patchwork-id: 307843
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 279/300] perf machine: Fill map_symbol->maps in append_inlines() to fix segfault
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 77b91c1a525d84cb560a4baef6f5f548b5c23f80
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Nov 29 15:47:51 2019 -0300

description
===========
I forgot to fill in the map_symbol->maps field in append_inlines() which
then makes code down the line segfault when trying to deref it.

It doesn't make any sense to have an addr_location with its 'map' member
not NULL while its 'maps' is NULL, after all al->maps is where al->map
is in.

It is done that way so that we don't have to have in each 'struct map' a
pointer to the 'struct maps' it is in, as we had in the past when we
would have 'map->mg', before 'struct maps' was combined with 'struct
map_groups', because there was always a one-to-one relationship for
these structs.

This fixes a segfault when processing DWARF callgraphs in 'perf report'.

Fixes: 08f6680e627e ("perf tools: Add a 'struct map_groups' pointer to 'struct map_symbol'")
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>