[tools] libtraceevent: Fix lib installation with O=
Message-id: <20200514101858.720-282-mpetlan@redhat.com>
Patchwork-id: 307845
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 281/300] libtraceevent: Fix lib installation with O=
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 587db8ebdac2c5eb3a8851e16b26f2e2711ab797
Author: Sudip Mukherjee <sudipm.mukherjee@gmail.com>
Date: Fri Nov 15 11:36:09 2019 +0000

description
===========
When we use 'O=' with make to build libtraceevent in a separate folder
it fails to install libtraceevent.a and libtraceevent.so.1.1.0 with the
error:

  INSTALL  /home/sudip/linux/obj-trace/libtraceevent.a
  INSTALL  /home/sudip/linux/obj-trace/libtraceevent.so.1.1.0

  cp: cannot stat 'libtraceevent.a': No such file or directory
  Makefile:225: recipe for target 'install_lib' failed
  make: *** [install_lib] Error 1

I used the command:

  make O=../../../obj-trace DESTDIR=~/test prefix==/usr  install

It turns out libtraceevent Makefile, even though it builds in a separate
folder, searches for libtraceevent.a and libtraceevent.so.1.1.0 in its
source folder.

So, add the 'OUTPUT' prefix to the source path so that 'make' looks for
the files in the correct place.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>