[tools] libtraceevent: Copy pkg-config file to output folder when using O=
Message-id: <20200514101858.720-283-mpetlan@redhat.com>
Patchwork-id: 307846
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 282/300] libtraceevent: Copy pkg-config file to output folder when using O=
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit 15b3904f8e884e0d34d5f09906cf6526d0b889a2
Author: Sudip Mukherjee <sudipm.mukherjee@gmail.com>
Date: Fri Nov 15 11:36:10 2019 +0000

description
===========
When we use 'O=' with make to build libtraceevent in a separate folder
it still copies 'libtraceevent.pc' to its source folder. Modify the
Makefile so that it uses the output folder to copy the pkg-config file
and install from there.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>