[tools] perf map: Set kmap->kmaps backpointer for main kernel map chunks
Message-id: <20200514101858.720-295-mpetlan@redhat.com>
Patchwork-id: 307858
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 294/300] perf map: Set kmap->kmaps backpointer for main kernel map chunks
Bugzilla: 1818838
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818838
https://bugzilla.redhat.com/show_bug.cgi?id=1818838

upstream
========
commit a75af86b6f344f99825cc894596804a91a2ee9cf
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Dec 18 15:23:14 2019 -0300

description
===========
When a map is create to represent the main kernel area (vmlinux) with
map__new2() we allocate an extra area to store a pointer to the 'struct
maps' for the kernel maps, so that we can access that struct when
loading ELF files or kallsyms, as we will need to split it in multiple
maps, one per kernel module or ELF section (such as ".init.text").

So when map->dso->kernel is non-zero, it is expected that
map__kmap(map)->kmaps to be set to the tree of kernel maps (modules,
chunks of the main kernel, bpf progs put in place via
PERF_RECORD_KSYMBOL, the main kernel).

This was not the case when we were splitting the main kernel into chunks
for its ELF sections, which ended up making 'perf report --children'
processing a perf.data file with callchains to trip on
__map__is_kernel(), when we press ENTER to see the popup menu for main
histogram entries that starts at a symbol in the ".init.text" ELF
section, e.g.:

-    8.83%     0.00%  swapper     [kernel.vmlinux].init.text  [k] start_kernel
     start_kernel
     cpu_startup_entry
     do_idle
     cpuidle_enter
     cpuidle_enter_state
     intel_idle

Fix it.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>