[netdrv] ice: Fix switch between FW and SW LLDP
Message-id: <21f5a22b1a889f70a3e0976d72182ce9929dae2a.1589816732.git.jtoppins@redhat.com>
Patchwork-id: 310413
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL 8.3 082/148] ice: Fix switch between FW and SW LLDP
Bugzilla: 1781388
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When switching between FW and SW LLDP mode, the
number of configured TLV apps in the driver's
DCB configuration is getting out of synch with
what lldpad thinks is configured.  This is causing
a problem when shutting down lldpad.  The cleanup
is trying to delete TLV apps that are not defined
in the kernel.

Since the driver is keeping an accurate account
of the apps defined, use the drivers number of
apps to determine if there is an app to delete.
If the number of apps is <= 1, then do not
attempt to delete.

Signed-off-by: Dave Ertman <david.m.ertman@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 53977ee47410885e7d4eee87d2c811a48a275150)
Bugzilla: 1781388
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28604488
Tested: posted for partner to test as well as various tests internally
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>