[sound] ALSA: ASoC: ti: davinci-mcasp: Improve the sysclk selection
Message-id: <15898774120838792-544-patchset-jkysela@redhat.com>
Patchwork-id: 311138
Patchwork-instance: patchwork
O-Subject: [RHEL 8.3 PATCH v2 0543/1765] ALSA: ASoC: ti: davinci-mcasp: Improve the sysclk selection
Bugzilla: 1797509
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When McASP is master the bclk can be generated from two main source:
AUXCLK: functional clock for McASP or
AHCLK: from external source or internal mux in dra7x family

With this patch it is possible to select between the two source. The patch
is not breaking existing machine drivers since historically the clk_id was
ignored and left as 0 in all cases.

When output clock is configured - which can be only the AHCLK, we select
the AUXCLK as source for the internal HCLK. In this case the HCLK rate is
the same as the output clock.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Link: https://lore.kernel.org/r/20191204192005.31210-1-peter.ujfalusi@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 253f584a0699d12a90bde9d524d499a921cc7827)
Bugzilla: 1797509
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>