[sound] ALSA: ASoC: hdmi-codec: re-introduce mutex locking again
Message-id: <15898774120838792-548-patchset-jkysela@redhat.com>
Patchwork-id: 311375
Patchwork-instance: patchwork
O-Subject: [RHEL 8.3 PATCH v2 0547/1765] ALSA: ASoC: hdmi-codec: re-introduce mutex locking again
Bugzilla: 1797509
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The dai codec needs to ensure that on one dai is used at any time.
This is currently protected by bit atomic operation. With this change,
it done with a mutex instead.

This change is not about functionality or efficiency. It is done with
the hope that it help maintainability in the future.

Suggested-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Link: https://lore.kernel.org/r/20191206103542.485224-1-jbrunet@baylibre.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 1442842952ccfe4178602c7a8459ac2ecb9d9e1a)
Bugzilla: 1797509
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>