[sound] ALSA: ASoC: tas2770: Fix snd_soc_update_bits error handling
Message-id: <15898774120838792-852-patchset-jkysela@redhat.com>
Patchwork-id: 311362
Patchwork-instance: patchwork
O-Subject: [RHEL 8.3 PATCH v2 0851/1765] ALSA: ASoC: tas2770: Fix snd_soc_update_bits error handling
Bugzilla: 1797509
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

According the documentation for snd_soc_update_bits the API
will return a 1 if the update was successful with a value change,
a 0 if the update was successful with no value change or a negative
if the command just failed.

So the value of return in the driver needs to be checked for being less
then 0 or the caller may indicate failure when the value actually
changed.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Link: https://lore.kernel.org/r/20191007171157.17813-1-dmurphy@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit edd6d5393206ec9bfc74776f6f20a57f11327e1b)
Bugzilla: 1797509
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>