[sound] ALSA: usb-audio: Add delayed_register option
Message-id: <15898774120838792-1453-patchset-jkysela@redhat.com>
Patchwork-id: 312195
Patchwork-instance: patchwork
O-Subject: [RHEL 8.3 PATCH v2 1452/1765] ALSA: usb-audio: Add delayed_register option
Bugzilla: 1797509
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Add a new option for specifying the quirk for delayed registration of
the certain device.  A list of devices can be passed in a form
	ID:IFACE,ID:IFACE,ID:IFACE,....
where ID is the 32bit hex number combo of vendor and device IDs and
IFACE is the interface number to trigger the register.

When a matching device is probed, the card registration is delayed
until the given interface is probed.  It's needed for syncing the
registration until the last interface when multiple interfaces are
provided for the same card.

Link: https://lore.kernel.org/r/20200325103322.2508-3-tiwai@suse.de
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit b70038ef4feae9c5c9d01de4cd799ae52ccafc08)
Bugzilla: 1797509
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>