[sound] ALSA: soundwire: ignore uniqueID when irrelevant
Message-id: <15898774120838792-1708-patchset-jkysela@redhat.com>
Patchwork-id: 312152
Patchwork-instance: patchwork
O-Subject: [RHEL 8.3 PATCH v2 1707/1765] ALSA: soundwire: ignore uniqueID when irrelevant
Bugzilla: 1797509
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The uniqueID is useful when there are two or more devices of the same
type (identical manufacturer ID, part ID) on the same link.

When there is a single device of a given type on a link, its uniqueID
is irrelevant. It's not uncommon on actual platforms to see variations
of the uniqueID, or differences between devID registers and ACPI _ADR
fields.

This patch suggests a filter on startup to identify 'single' devices
and tag them accordingly. The uniqueID is then not used for the probe,
and the device name omits the uniqueID as well.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20191022234808.17432-4-pierre-louis.bossart@linux.intel.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 2e8c4ad1f04413a4a67ef10746a7566007d2ed55)
Bugzilla: 1797509
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>