[fs] ceph: register MDS request with dir inode from the start
Message-id: <20200519180620.418410-65-jlayton@redhat.com>
Patchwork-id: 312684
Patchwork-instance: patchwork
O-Subject: [RHEL8.3.0 PATCH BZ#1827767 064/119] ceph: register MDS request with dir inode from the start
Bugzilla: 1827767
RH-Acked-by: Xiubo Li <xiubli@redhat.com>
RH-Acked-by: Zheng Yan <zyan@redhat.com>

From: Jeff Layton <jlayton@kernel.org>

When the unsafe reply to a request comes in, the request is put on the
r_unsafe_dir inode's list. In future patches, we're going to need to
wait on requests that may not have gotten an unsafe reply yet.

Change __register_request to put the entry on the dir inode's list when
the pointer is set in the request, and don't check the
CEPH_MDS_R_GOT_UNSAFE flag when unregistering it.

The only place that uses this list today is fsync codepath, and with
the coming changes, we'll want to wait on all operations whether it has
gotten an unsafe reply or not.

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: "Yan, Zheng" <zyan@redhat.com>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 3db0a2fc5668c6dc60218e0f55d2443bcd726aa0)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>