[fs] ceph: add flag to designate that a request is asynchronous
Message-id: <20200519180620.418410-87-jlayton@redhat.com>
Patchwork-id: 312724
Patchwork-instance: patchwork
O-Subject: [RHEL8.3.0 PATCH BZ#1827767 086/119] ceph: add flag to designate that a request is asynchronous
Bugzilla: 1827767
RH-Acked-by: Xiubo Li <xiubli@redhat.com>
RH-Acked-by: Zheng Yan <zyan@redhat.com>

From: Jeff Layton <jlayton@kernel.org>

...and ensure that such requests are never queued. The MDS has need to
know that a request is asynchronous so add flags and proper
infrastructure for that.

Also, delegated inode numbers and directory caps are associated with the
session, so ensure that async requests are always transmitted on the
first attempt and are never queued to wait for session reestablishment.

If it does end up looking like we'll need to queue the request, then
have it return -EJUKEBOX so the caller can reattempt with a synchronous
request.

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: "Yan, Zheng" <zyan@redhat.com>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 3bb48b4142bbf72045af5ebe72e65ccff6d02680)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>