[fs] ceph: cache layout in parent dir on first sync create
Message-id: <20200519180620.418410-97-jlayton@redhat.com>
Patchwork-id: 312721
Patchwork-instance: patchwork
O-Subject: [RHEL8.3.0 PATCH BZ#1827767 096/119] ceph: cache layout in parent dir on first sync create
Bugzilla: 1827767
RH-Acked-by: Xiubo Li <xiubli@redhat.com>
RH-Acked-by: Zheng Yan <zyan@redhat.com>

From: Jeff Layton <jlayton@kernel.org>

If a create is done, then typically we'll end up writing to the file
soon afterward. We don't want to wait for the reply before doing that
when doing an async create, so that means we need the layout for the
new file before we've gotten the response from the MDS.

All files created in a directory will initially inherit the same layout,
so copy off the requisite info from the first synchronous create in the
directory, and save it in a new i_cached_layout field. Zero out the
layout when we lose Dc caps in the dir.

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: "Yan, Zheng" <zyan@redhat.com>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 785892fe88f693028a996d1cb363d60609848e5e)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>

Conflicts:
	fs/ceph/file.c - RHEL8 uses FILE_CREATED, not FMODE_CREATED