[idle] intel_idle: Allow ACPI _CST to be used for selected known processors
Message-id: <20200519182515.19628.91856.sendpatchset@sbest.bos.csb>
Patchwork-id: 312749
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.3 BZ1485545 3/6] intel_idle: Allow ACPI _CST to be used for selected known processors
Bugzilla: 1485545
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1485545
Upstream:
commit bff8e60a86f4960133f90ad9add9adeb082b8154

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=bff8e60a86f4960133f90ad9add9adeb082b8154

Description:
Update the intel_idle driver to get the C-states information from ACPI
_CST in some cases in which the processor is known to the driver, as long as
that information is available and the new use_acpi flag is set in the
profile of the processor in question.

In the cases when there is a specific table of C-states for the given
processor in the driver, that table is used as the primary source of
information on the available C-states, but if ACPI _CST is present,
the C-states that are not listed by it will not be enabled by default
(they still can be enabled later by user space via sysfs, though).

The new CPUIDLE_FLAG_ALWAYS_ENABLE flag can be used for marking
C-states that should be enabled by default even if they are not
listed by ACPI _CST.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>